refactor: 디렉토리명 리팩토링
