+++ lib/screens/bottom_navigation/more/more_page_titles/more_web_navigation_tile.dart
@@ -5,12 +5,12 @@
  * For full license text, see the LICENSE file in the root directory or at
  * http://www.apache.org/licenses/
  * Author: junho Kim
- * Latest Updated Date: 2025-03-01
+ * Latest Updated Date: 2025-07-06
  */
 import 'package:flutter/material.dart';
 import 'package:inha_notice/fonts/font.dart';
 import 'package:inha_notice/themes/theme.dart';
-import 'package:inha_notice/widgets/navigation/web_navigator.dart';
+import 'package:inha_notice/widgets/webview//web_navigator.dart';
 
 /// **MoreWebNavigationTile**
 /// 이 클래스는 더보기 페이지의 웹 네이게이션 타일을 정의하는 클래스입니다.
