+++ lib/screens/bottom_navigation/more/notification_setting/categories/education_notification_category.dart
@@ -5,14 +5,14 @@
  * For full license text, see the LICENSE file in the root directory or at
  * http://www.apache.org/licenses/
  * Author: junho Kim
- * Latest Updated Date: 2025-05-19
+ * Latest Updated Date: 2025-07-06
  */
 
 import 'package:flutter/material.dart';
 import 'package:inha_notice/constants/shared_pref_keys/shared_pref_keys.dart';
 import 'package:inha_notice/screens/bottom_navigation/more/notification_setting/notification_tile.dart';
 import 'package:inha_notice/themes/theme.dart';
-import 'package:inha_notice/widgets/texts/bold_title_text.dart';
+import 'package:inha_notice/widgets/typography/bold_title_text.dart';
 
 class EducationNotificationCategory extends StatefulWidget {
   const EducationNotificationCategory({super.key});
