+++ lib/screens/bottom_navigation/more/notification_setting/categories/undergraduate_schedule_notification_category.dart
@@ -1,8 +1,18 @@
+/*
+ * This is file of the project inha_notice
+ * Licensed under the Apache License 2.0.
+ * Copyright (c) 2025 INGONG
+ * For full license text, see the LICENSE file in the root directory or at
+ * http://www.apache.org/licenses/
+ * Author: junho Kim
+ * Latest Updated Date: 2025-07-06
+ */
+
 import 'package:flutter/material.dart';
 import 'package:inha_notice/constants/shared_pref_keys/shared_pref_keys.dart';
 import 'package:inha_notice/screens/bottom_navigation/more/notification_setting/notification_tile.dart';
 import 'package:inha_notice/themes/theme.dart';
-import 'package:inha_notice/widgets/texts/bold_title_text.dart';
+import 'package:inha_notice/widgets/typography/bold_title_text.dart';
 
 class UndergraduateScheduleNotificationCategory extends StatefulWidget {
   const UndergraduateScheduleNotificationCategory({super.key});
