+++ lib/widgets/webview/in_app_web_page.dart
@@ -1,12 +1,12 @@
 /*
-* This is file of the project inha_notice
-* Licensed under the Apache License 2.0.
-* Copyright (c) 2025 INGONG
-* For full license text, see the LICENSE file in the root directory or at
-* http://www.apache.org/licenses/
-* Author: junho Kim
-* Latest Updated Date: 2025-02-10
-*/
+ * This is file of the project inha_notice
+ * Licensed under the Apache License 2.0.
+ * Copyright (c) 2025 INGONG
+ * For full license text, see the LICENSE file in the root directory or at
+ * http://www.apache.org/licenses/
+ * Author: junho Kim
+ * Latest Updated Date: 2025-07-06
+ */
 import 'dart:io';
 
 import 'package:flutter/material.dart';
